teste-whats/api/pyproject.toml
[project]
name = "teste-whats-api"
version = "0.1.0"
description = "API FastAPI com MongoDB para a frente Teste Whats."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "pymongo>=4.8.0,<5.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
